requirements.txt
Django==2.2
requests==2.21.0
boto3==1.9.208
botocore==1.12.208
djangorestframework==3.10.2
django_storages==1.7.1
django-cors-headers==3.0.2
mysqlclient
setuptools-rust==0.12.1
django-environ==0.4.5
django-mail-templated==2.6.5
djangorestframework-jwt==1.11.0
bcrypt==3.1.7
djangorestframework-simplejwt==4.3.0
Pillow
numpy==1.19.5
crc16==0.1.1
geopy==2.0.0
pyqrcode==1.2.1
pypng==0.0.20
cryptography==3.2.1
channels~=2.4.0
asgiref~=3.2.6
httplib2==0.19.1
ratelimit==2.2.1
backoff==1.10.0
django-bootstrap4
xhtml2pdf
pytesseract==0.3.5
tensorflow
tensorflow-estimator
xgboost==1.4.2
sinling==0.3.6
simplejson==3.16.0
django-matplotlib==0.1
djongo==1.3.2
oauthlib==3.1.0
pandas==1.3.0
scikit-image==0.18.1
scikit-learn==0.24.2
imutils==0.5.4
joblib==1.0.1
jsonschema==3.2.0
keras-Preprocessing==1.1.2
opencv-python==4.4.0.44
packaging==20.9
pymongo
regex==2021.7.6
sklearn-crfsuite==0.3.6
sqlparse==0.2.4
wheel
django-bootstrap4
dnspython
regex == 2021.10.8

